requirements.txt
googlemaps
beautifulsoup4
requests
pyodbc
sqlalchemy
python-dotenv
typing
black
pylint
pytest
pytest-mock